tools/content_email_agent/requirements.txt
openai>=1.57.0
python-dotenv>=1.0.1
pypdf>=5.0.0
python-docx>=1.1.2
openpyxl>=3.1.5
